Move esp8266 specific files into example folder
